Move files; change Documentation Files collection path
